third_party/blink/web_tests/external/wpt/css/css-tables/html-display-table.html

<!DOCTYPE html>
<html style="margin:0;padding:0">
<title>&lt;html display:table&gt;</title>
<link rel="author" href="mailto:atotic@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://www.w3.org/TR/css-overflow-3/#scrollable">
<meta name="assert" content="<html display:table> does not fill the viewport">
<style>
  html {
    display: table;
    border: 10px solid green;
    border-spacing: 0;
    padding: 0;
    margin: 0;
  }
  body {
    padding: 0;
    margin: 0;
  }
</style>
<div style="width:200px;height:300px;background:yellow;display:inline-block;"></div><div style="width:80px;height:300px;background:yellow;display:inline-block;"></div>
<script>
  test( _=> {
    assert_equals(document.querySelector("html").offsetWidth, 300);
  }, "<html display:table> width is table's max width");
</script>
